+++ packages/xchain-thorchain-query/src/utils/swap.ts
@@ -35,6 +35,14 @@ import { LiquidityPool } from '../liquidity-pool'
 import { ThorchainCache } from '../thorchain-cache'
 import { SwapOutput } from '../types'
 
+const getBaseAmountWithDiffDecimals = (inputAmount: CryptoAmount, outDecimals: number): BigNumber => {
+  const inDecimals = inputAmount.baseAmount.decimal
+  let baseAmountOut = inputAmount.baseAmount.amount()
+  const adjustDecimals = outDecimals - inDecimals
+  baseAmountOut = baseAmountOut.times(10 ** adjustDecimals)
+  return baseAmount(baseAmountOut, outDecimals).amount()
+}
+
 /**
  *
  * @param inputAmount - amount to swap
@@ -46,16 +54,20 @@ export const getSwapFee = (inputAmount: CryptoAmount, pool: LiquidityPool, toRun
   // formula: (x * x * Y) / (x + X) ^ 2
   // const isInputRune = isAssetRuneNative(inputAmount.asset)
 
-  const x = inputAmount.baseAmount.amount()
+  const x = getBaseAmountWithDiffDecimals(inputAmount, 8)
   const X = toRune ? pool.assetBalance.amount() : pool.runeBalance.amount() // input is asset if toRune
   const Y = toRune ? pool.runeBalance.amount() : pool.assetBalance.amount() // output is rune if toRune
   const units = toRune ? AssetRuneNative : pool.asset
-  const decimals = toRune || !pool.decimals ? 8 : pool.decimals
   const numerator = x.times(x).multipliedBy(Y)
   const denominator = x.plus(X).pow(2)
   const result = numerator.div(denominator)
-  const swapFee = new CryptoAmount(baseAmount(result, decimals), units)
-  // console.log(` swapFee ${swapFee.assetAmountFixedString()} ${assetToString(units)}`)
+
+  const eightDecimalResult = new CryptoAmount(baseAmount(result), units)
+
+  const decimals = toRune ? 8 : inputAmount.baseAmount.decimal
+  const baseOut = getBaseAmountWithDiffDecimals(eightDecimalResult, decimals)
+  const swapFee = new CryptoAmount(baseAmount(baseOut, decimals), units)
+  //console.log(` swapFee ${swapFee.assetAmountFixedString()} `)
   return swapFee
 }
 
@@ -69,7 +81,7 @@ export const getSwapFee = (inputAmount: CryptoAmount, pool: LiquidityPool, toRun
  */
 export const getSwapSlip = (inputAmount: CryptoAmount, pool: LiquidityPool, toRune: boolean): BigNumber => {
   // formula: (x) / (x + X)
-  const x = inputAmount.baseAmount.amount()
+  const x = getBaseAmountWithDiffDecimals(inputAmount, 8)
   const X = toRune ? pool.assetBalance.amount() : pool.runeBalance.amount() // input is asset if toRune
   const result = x.div(x.plus(X))
   return new BigNumber(result)
@@ -84,16 +96,21 @@ export const getSwapSlip = (inputAmount: CryptoAmount, pool: LiquidityPool, toRu
  */
 export const getSwapOutput = (inputAmount: CryptoAmount, pool: LiquidityPool, toRune: boolean): CryptoAmount => {
   // formula: (x * X * Y) / (x + X) ^ 2
-  const x = inputAmount.baseAmount.amount()
+  const x = getBaseAmountWithDiffDecimals(inputAmount, 8)
   const X = toRune ? pool.assetBalance.amount() : pool.runeBalance.amount() // input is asset if toRune
   const Y = toRune ? pool.runeBalance.amount() : pool.assetBalance.amount() // output is rune if toRune
 
   const units = toRune ? AssetRuneNative : pool.asset
-  const decimals = toRune || !pool.decimals ? 8 : pool.decimals
+  // const decimals = toRune || !pool.decimals ? 8 : pool.decimals
   const numerator = x.times(X).times(Y)
   const denominator = x.plus(X).pow(2)
   const result = numerator.div(denominator)
-  return new CryptoAmount(baseAmount(result, decimals), units)
+
+  const eightDecimalResult = new CryptoAmount(baseAmount(result), units)
+
+  const decimals = toRune ? 8 : inputAmount.baseAmount.decimal
+  const baseOut = getBaseAmountWithDiffDecimals(eightDecimalResult, decimals)
+  return new CryptoAmount(baseAmount(baseOut, decimals), units)
 }
 
 export const getDoubleSwapOutput = (
@@ -214,9 +231,9 @@ export const calcNetworkFee = (asset: Asset, gasRate: BigNumber): CryptoAmount =
       const gasRateinAVAXGwei = gasRate
       const gasRateinAVAXWei = baseAmount(gasRateinAVAXGwei.multipliedBy(10 ** 9), 18)
       if (eqAsset(asset, AssetAVAX)) {
-        return new CryptoAmount(gasRateinAVAXWei.times(21000), AssetETH)
+        return new CryptoAmount(gasRateinAVAXWei.times(21000), AssetAVAX)
       } else {
-        return new CryptoAmount(gasRateinAVAXWei.times(70000), AssetETH)
+        return new CryptoAmount(gasRateinAVAXWei.times(70000), AssetAVAX)
       }
       break
     case Chain.Terra: