+++ packages/xchain-thorchain-query/CHANGELOG.md
@@ -1,3 +1,16 @@
+# v0.0.1.0-beta (2022-09-15)
+
+## Fix
+
+- fixed math calcs to account for decimals
+
+# v0.0.1.0-alpha1 (2022-09-07)
+
+## Fix
+
+- Updated EstimateSwapParam type
+- Edited function `estimateSwap()` to include default parameters
+
 # v0.0.1.0-alpha (2022-08-26)
 
 ## Module Created